+++ code/bulkloader/ListBulkLoader.php
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * Peforms bulk loading, but works with a given DataList.
+ */
+class ListBulkLoader extends BetterBulkLoader {
+
+	/**
+	 * The list to insert new DataObjects into.
+	 * @var DataList
+	 */
+	protected $list;
+
+	public function __construct(DataList $list) {
+		$this->list = $list;
+
+		parent::__construct($this->list->dataClass());
+	}
+
+	public function setList(DataList $list){
+		$this->list = $source;
+
+		return $this;
+	}
+
+	/**
+	 * Add records to the list.
+	 */
+	protected function processAll($filepath, $preview = false) {
+		$iterator = $this->getSource()->getIterator();
+		$results = new BetterBulkLoader_Result();
+
+		foreach($iterator as $record) {
+			if($id = $this->processRecord($record, $this->columnMap, $results, $preview)){
+				$this->list->add($id);
+			}
+		}
+		
+		
+
+		return $results;
+	}
+
+	public function deleteExistingRecords(){
+		$this->list->removeAll();
+	}
+
+}